se cambia el tipo del dato double

+++ store-api/src/main/java/com/company/storeapi/core/mapper/TicketMapper.java
@@ -80,33 +80,33 @@ public abstract class TicketMapper {
         ticket.setPaymentType(requestAddTicketDTO.getPaymentType());
 
         ticket.setTicketStatus(TicketStatus.PAYED);
-        ticket.setOutstandingBalance((double) 0);
+        ticket.setOutstandingBalance(0);
 
-        Double getTicketCostWithoutIVA = (IVA.IVA19 * order.getTotalOrder()) / IVA.PORCENTAJE;
+        double getTicketCostWithoutIVA = (IVA.IVA19 * order.getTotalOrder()) / IVA.PORCENTAJE;
         ticket.setTicketCost(order.getTotalOrder());
         ticket.setTicketCostWithoutIVA(getTicketCostWithoutIVA);
 
         ticket.setCashPayment(order.getTotalOrder());
-        ticket.setTransactionPayment((double) 0);
-        ticket.setCreditPayment((double) 0);
+        ticket.setTransactionPayment(0);
+        ticket.setCreditPayment(0);
 
         if (ticket.getPaymentType() == PaymentType.TRANSACTION) {
             ticket.setTransactionPayment(order.getTotalOrder());
-            ticket.setCashPayment((double) 0);
-            ticket.setCreditPayment((double) 0);
+            ticket.setCashPayment(0);
+            ticket.setCreditPayment(0);
         }
 
         if (ticket.getPaymentType() == PaymentType.CREDIT) {
-            ticket.setTransactionPayment((double) 0);
-            ticket.setCashPayment((double) 0);
+            ticket.setTransactionPayment(0);
+            ticket.setCashPayment(0);
             ticket.setCreditPayment(order.getTotalOrder());
 
             Set<CreditCapital> creditCapitals = new LinkedHashSet<>();
             CreditCapital creditCapital = new CreditCapital();
             creditCapital.setCashCreditCapital(requestAddTicketDTO.getCreditCapital());
-            creditCapital.setTransactionCreditCapital((double) 0);
+            creditCapital.setTransactionCreditCapital(0);
             if (requestAddTicketDTO.getCreditPaymentType() == PaymentType.TRANSACTION) {
-                creditCapital.setCashCreditCapital((double) 0);
+                creditCapital.setCashCreditCapital(0);
                 creditCapital.setTransactionCreditCapital(requestAddTicketDTO.getCreditCapital());
             }
             creditCapital.setCreatAt(new Date());
